rename: Card 컴포넌트 명을CardPreview로 변경
